Allow locally defined constant to override filter.
While working locally, I ran into a (rare) situation which I needed to set `JETPACK_DEV_DEBUG` to `false` momentarily. However, because of the filter set in Airplane Mode, I also had to set it to Inactive (which is not really an issue).

That being said, it seems the locally defined constant _should_ override the filter in this case and this commit addresses that issue.